Create torque_vs_angles_1joint
I plotted the vector position,torque vs phi and rho vs phi. The maximum torque is reached around 30 deg whereas the maximum rho (moment arm) is reached at 60 deg. I would expected the maximum arm and the maximum momentum for the same deg but I interpreted the datas in this way: for 30 deg there is the better combination between the force and L to generate torque; the best relative position between F and L is reached at 60 deg but in this case F has less magnitude